fixed bug in convert file

+++ site/convert_notebooks.py
@@ -2,9 +2,15 @@ import os,glob
 
 nbc_str_tpl = 'jupyter nbconvert %s --to markdown --NbConvertApp.output_files_dir=.'
 
+
+
 base_dir = os.getcwd()
 post_dirs = glob.glob('content/entry/*')
 post_dirs_fp = [base_dir + '/' + p for p in post_dirs]
+
+nb_dir = base_dir + '/../notebooks'
+
+
 
 top_lines = \
 """
@@ -39,7 +45,7 @@ index_txt = \
 """
 
 +++
-title = "Making Word Clouds"  # Add a page title.
+title = "Title %s"  # Add a page title.
 summary = "Hello!"  # Add a page description.
 date = 2019-01-01T00:00:00  # Add today's date.
 type = "widget_page"  # Page type is a Widget Page
@@ -47,22 +53,47 @@ type = "widget_page"  # Page type is a Widget Page
 
 """
 
-for p in post_dirs_fp:
+os.chdir(nb_dir)
+nbs = glob.glob('*.ipynb')
 
-    os.chdir(p)
-    ipynb_file = glob.glob('*.ipynb')[0] # assume there is only one nb
-    nbc_str = nbc_str_tpl % ipynb_file 
-    os.system(nbc_str)
+entries_dir = base_dir + '/content/entry' 
+
+print('notebooks: %s' %nbs)
+
+
+for nb_it,nb in enumerate(nbs):
+
+    nb_full = nb_dir + '/' + nb
+    
+    nb_name = nb.replace('.ipynb', '')
+        
+    new_dir = entries_dir + '/%s' % nb_name
+    print('going to new dir: %s' %new_dir)
+    if not os.path.isdir(new_dir): os.makedirs(new_dir)
+    os.chdir(new_dir)
+    
+    print('copying file %s to .' %nb_full)
+    os.system('cp %s .'  %nb_full)
+
+    if os.path.isfile(nb):
+
+      print('running nbconvert for %s' %nb)
+
+      nbc_str = nbc_str_tpl % nb 
+      os.system(nbc_str)
    
     
-    md_file = ipynb_file.replace('.ipynb', '.md')
-    md_str = top_lines + open(md_file, 'r').read()
-    open(md_file, 'w+').write(md_str)
+      md_file = nb_name + '.md'
+      md_str = top_lines + open(md_file, 'r').read()
+      open(md_file, 'w+').write(md_str)
 
-    index_file = 'index.md'
-    open(index_file, 'w+').write(index_txt)
-    
-    
+      index_file = 'index.md'
+      open(index_file, 'w+').write(index_txt %nb_it)    
+   
+    else:
+
+      print('%s - file not found' %nb)
+
 os.chdir(base_dir)
 
 #os.system('cat %s.md %s.md | tee -a index.md')
